Moved helper normalising functions to `sunpy.image.util`
